move salsa memory dump to `ProjectDatabase`
